Trying to configure AndroidX

+++ app/src/main/java/com/themovielist/model/MovieSizeModel.kt
@@ -4,14 +4,14 @@ import android.os.Parcel
 import android.os.Parcelable
 import com.themovielist.enums.ImageSizeEnum
 
-data class MovieSizeModel constructor(val size: Int, @ImageSizeEnum.ImageSize val sizeType: Int) : Parcelable {
+data class MovieSizeModel constructor(val size: Int, @ImageSizeEnum.ImageSize val sizeType: Long) : Parcelable {
     constructor(parcel: Parcel) : this(
             parcel.readInt(),
-            parcel.readInt())
+            parcel.readLong())
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeInt(size)
-        parcel.writeInt(sizeType)
+        parcel.writeLong(sizeType)
     }
 
     override fun describeContents(): Int {
